feat: update auth folder
