+++ src/schema/resolvers/Auth/Mutation/index.ts
@@ -1,0 +1,5 @@
+import login from './login'
+
+export default {
+  login,
+}
